refactor(tailwind-variants): consolidate and reorganize `createTV` and `defaultConfig`
Moved `createTV` and `defaultConfig` logic directly into `tv.ts` for better modularity and maintainability. Simplified exports by reorganizing and centralizing them, reducing redundancy across modules. This refactor enhances the overall structure and clarity of the `tailwind-variants` package.
