--- packages/tailwind-variants/src/core/config.ts
@@ -1,6 +1,0 @@
-import type { Config } from "@/types";
-
-export const defaultConfig: Config = {
-  twMerge: true,
-  twMergeConfig: {},
-};
