+++ packages/tailwind-variants/src/core/index.ts
@@ -1,4 +1,1 @@
-export { defaultConfig } from "@/core/config";
-export { createTV } from "@/core/create-tv";
-export { tv } from "@/core/tv";
-export type { Config, TV } from "@/types";
+export { createTV, defaultConfig, tv } from "@/core/tv";
